backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Abhishek Tiwari",email = "abhishek.tiwari@skill-mine.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
